Hardcode proxy path 🎨
