+++ .env.example
@@ -70,9 +70,6 @@ STORAGE_BUCKET=
 STORAGE_PROXY=local
 STORAGE_PROXY_PRIVATE_KEY=
 
-# Local settings
-STORAGE_PROXY_PATH=/storage
-
 # Cloudfront settings
 STORAGE_PROXY_HOST=
 STORAGE_KEY_PAIR_ID=
