+++ server/router.js
@@ -3,8 +3,6 @@
 const { auth: authConfig } = require('../config/server');
 const { authenticate } = require('./shared/auth');
 const express = require('express');
-const { getFile } = require('./shared/storage/proxy/mw');
-const proxy = require('./shared/storage/proxy');
 const repository = require('./repository');
 const storage = require('./shared/storage/storage.router');
 const tag = require('./tag');
@@ -14,9 +12,6 @@ const router = express.Router();
 router.use(processBody);
 
 // Public routes:
-if (proxy.isSelfHosted) {
-  router.use(proxy.provider.path, getFile);
-}
 router.use(user.path, user.router);
 
 // SSO routes:
