+++ server/shared/storage/storage.router.js
@@ -3,11 +3,10 @@
 const ctrl = require('./storage.controller');
 const multer = require('multer');
 const router = require('express').Router();
-const { setSignedCookies } = require('./proxy/mw');
 const upload = multer({ storage: multer.memoryStorage() });
 
 router
-  .get('/set-cookies', setSignedCookies)
+  .get('/set-cookies', ctrl.setSignedCookies)
   .post('/', upload.single('file'), ctrl.upload);
 
 module.exports = {
